+++ system/src/Grav/Framework/File/Formatter/FormatterInterface.php
@@ -1,7 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 /**
- * @package    Grav\Framework\File\Formatter
+ * @package    Grav\Framework\File
  *
  * @copyright  Copyright (C) 2015 - 2018 Trilby Media, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
@@ -9,7 +11,7 @@
 
 namespace Grav\Framework\File\Formatter;
 
-interface FormatterInterface
+interface FormatterInterface extends \Serializable
 {
     /**
      * Get default file extension from current formatter (with dot).
@@ -18,14 +20,14 @@ interface FormatterInterface
      *
      * @return string File extension (can be empty).
      */
-    public function getDefaultFileExtension();
+    public function getDefaultFileExtension(): string;
 
     /**
      * Get file extensions supported by current formatter (with dot).
      *
      * @return string[]
      */
-    public function getSupportedFileExtensions();
+    public function getSupportedFileExtensions(): array;
 
     /**
      * Encode data into a string.
@@ -33,13 +35,13 @@ interface FormatterInterface
      * @param array $data
      * @return string
      */
-    public function encode($data);
+    public function encode($data): string;
 
     /**
      * Decode a string into data.
      *
      * @param string $data
-     * @return array
+     * @return mixed
      */
     public function decode($data);
 }