+++ system/src/Grav/Framework/File/Formatter/YamlFormatter.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * @package    Grav\Framework\File\Formatter
  *
@@ -14,50 +16,63 @@ use Symfony\Component\Yaml\Exception\ParseException;
 use Symfony\Component\Yaml\Yaml as YamlParser;
 use RocketTheme\Toolbox\Compat\Yaml\Yaml as FallbackYamlParser;
 
-class YamlFormatter implements FormatterInterface
+class YamlFormatter extends AbstractFormatter
 {
-    /** @var array */
-    private $config;
-
     public function __construct(array $config = [])
     {
-        $this->config = $config + [
+        $config += [
             'file_extension' => '.yaml',
             'inline' => 5,
             'indent' => 2,
             'native' => true,
             'compat' => true
         ];
+
+        parent::__construct($config);
     }
 
     /**
-     * {@inheritdoc}
+     * @return int
      */
-    public function getDefaultFileExtension()
+    public function getInlineOption(): int
     {
-        $extensions = $this->getSupportedFileExtensions();
-
-        return (string) reset($extensions);
+        return $this->getConfig('inline');
     }
 
     /**
-     * {@inheritdoc}
+     * @return int
      */
-    public function getSupportedFileExtensions()
+    public function getIndentOption(): int
     {
-        return (array) $this->config['file_extension'];
+        return $this->getConfig('indent');
+    }
+
+    /**
+     * @return bool
+     */
+    public function useNativeDecoder(): bool
+    {
+        return $this->getConfig('native');
+    }
+
+    /**
+     * @return bool
+     */
+    public function useCompatibleDecoder(): bool
+    {
+        return $this->getConfig('compat');
     }
 
     /**
      * {@inheritdoc}
      */
-    public function encode($data, $inline = null, $indent = null)
+    public function encode($data, $inline = null, $indent = null): string
     {
         try {
-            return (string) YamlParser::dump(
+            return YamlParser::dump(
                 $data,
-                $inline ? (int) $inline : $this->config['inline'],
-                $indent ? (int) $indent : $this->config['indent'],
+                $inline ? (int) $inline : $this->getInlineOption(),
+                $indent ? (int) $indent : $this->getIndentOption(),
                 YamlParser::DUMP_EXCEPTION_ON_INVALID_TYPE
             );
         } catch (DumpException $e) {
@@ -68,10 +83,10 @@ class YamlFormatter implements FormatterInterface
     /**
      * {@inheritdoc}
      */
-    public function decode($data)
+    public function decode($data): array
     {
         // Try native PECL YAML PHP extension first if available.
-        if (\function_exists('yaml_parse') && $this->config['native']) {
+        if (\function_exists('yaml_parse') && $this->useNativeDecoder()) {
             // Safely decode YAML.
             $saved = @ini_get('yaml.decode_php');
             @ini_set('yaml.decode_php', 0);
@@ -79,14 +94,14 @@ class YamlFormatter implements FormatterInterface
             @ini_set('yaml.decode_php', $saved);
 
             if ($decoded !== false) {
-                return (array) $decoded;
+                return $decoded;
             }
         }
 
         try {
-            return (array) YamlParser::parse($data);
+            return YamlParser::parse($data);
         } catch (ParseException $e) {
-            if ($this->config['compat']) {
+            if ($this->useCompatibleDecoder()) {
                 return (array) FallbackYamlParser::parse($data);
             }